+++ src/main/java/dev/willyelton/crystal_tools/item/LevelableItem.java
@@ -2,9 +2,13 @@ package dev.willyelton.crystal_tools.item;
 
 import dev.willyelton.crystal_tools.config.CrystalToolsConfig;
 import dev.willyelton.crystal_tools.utils.NBTUtils;
+import net.minecraft.Util;
 import net.minecraft.core.BlockPos;
+import net.minecraft.network.chat.TextComponent;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.*;
 import net.minecraft.world.level.Level;
 
@@ -26,11 +30,11 @@ public interface LevelableItem {
 
 //    // From TierdItem.java
 
-    default void addExp(ItemStack tool, Level level, BlockPos blockPos) {
-        addExp(tool, level, blockPos, 1);
+    default void addExp(ItemStack tool, Level level, BlockPos blockPos, LivingEntity player) {
+        addExp(tool, level, blockPos, player, 1);
     }
 
-    default void addExp(ItemStack tool, Level level, BlockPos blockPos, int amount) {
+    default void addExp(ItemStack tool, Level level, BlockPos blockPos,LivingEntity livingEntity, int amount) {
         int newExperience = (int) NBTUtils.addValueToTag(tool, "experience", amount);
         int experienceCap = (int) NBTUtils.getFloatOrAddKey(tool, "experience_cap", CrystalToolsConfig.BASE_EXPERIENCE_CAP.get());
 
@@ -43,14 +47,20 @@ public interface LevelableItem {
         if (newExperience >= experienceCap) {
             // level up
             NBTUtils.addValueToTag(tool, "skill_points", 1);
-            // copied from LivingEntity item breaking sound
             // play level up sound
             level.playSound(null, blockPos.getX(), blockPos.getY(), blockPos.getZ(), SoundEvents.PLAYER_LEVELUP, SoundSource.NEUTRAL, 0.8F, 1.0F);
-            // TODO: Add chat message thing
-
+            if (livingEntity instanceof Player player) {
+                if (tool.getItem() instanceof LevelableItem item) {
+                    player.displayClientMessage(new TextComponent(tool.getItem().getName(tool).getString() + " Leveled Up (" + item.getSkillPoints(tool) + " Unspent Points)"), true);
+                }
+            }
             NBTUtils.setValue(tool, "experience", Math.max(0, newExperience - experienceCap));
             NBTUtils.setValue(tool, "experience_cap", (float) (experienceCap * CrystalToolsConfig.EXPERIENCE_MULTIPLIER.get()));
         }
+    }
+
+    default int getSkillPoints(ItemStack stack) {
+        return (int) NBTUtils.getFloatOrAddKey(stack, "skill_points");
     }
 
     String getItemType();